+++ src/views/payment-options-view.js
@@ -28,6 +28,7 @@ PaymentOptionsView.prototype._initialize = function () {
 };
 
 PaymentOptionsView.prototype._addPaymentOption = function (paymentOptionID) {
+  var paymentSource;
   var div = document.createElement('div');
   var html = paymentMethodOptionHTML;
   var clickHandler = function clickHandler() {
@@ -41,18 +42,24 @@ PaymentOptionsView.prototype._addPaymentOption = function (paymentOptionID) {
 
   switch (paymentOptionID) {
     case paymentOptionIDs.card:
+      paymentSource = this.strings.Card;
       html = html.replace(/@ICON/g, 'iconCardFront');
-      html = html.replace(/@OPTION_TITLE/g, this.strings.Card);
+      html = html.replace(/@OPTION_LABEL/g, this._generateOptionLabel(paymentSource));
+      html = html.replace(/@OPTION_TITLE/g, paymentSource);
       html = html.replace(/@CLASSNAME/g, 'braintree-icon--bordered');
       break;
     case paymentOptionIDs.paypal:
+      paymentSource = this.strings.PayPal;
       html = html.replace(/@ICON/g, 'logoPayPal');
+      html = html.replace(/@OPTION_LABEL/g, this._generateOptionLabel(this.strings.PayPal));
       html = html.replace(/@OPTION_TITLE/g, this.strings.PayPal);
       html = html.replace(/@CLASSNAME/g, '');
       break;
     case paymentOptionIDs.paypalCredit:
+      paymentSource = this.strings['PayPal Credit'];
       html = html.replace(/@ICON/g, 'logoPayPalCredit');
-      html = html.replace(/@OPTION_TITLE/g, this.strings['PayPal Credit']);
+      html = html.replace(/@OPTION_LABEL/g, this._generateOptionLabel(paymentSource));
+      html = html.replace(/@OPTION_TITLE/g, paymentSource);
       html = html.replace(/@CLASSNAME/g, '');
       break;
     default:
@@ -68,6 +75,10 @@ PaymentOptionsView.prototype._addPaymentOption = function (paymentOptionID) {
     div: div,
     clickHandler: clickHandler
   };
+};
+
+PaymentOptionsView.prototype._generateOptionLabel = function (paymentSourceString) {
+  return this.strings.payingWith.replace('{{paymentSource}}', paymentSourceString);
 };
 
 module.exports = PaymentOptionsView;