+++ CHANGELOG.md
@@ -3,9 +3,9 @@ CHANGELOG
 
 unreleased
 ----------
-
 - Use version 3.20.0 of braintree-web
 - Update browser-detection to v1.5.0
+- Add `aria-label` attribute to payment options
 - Update checkout.js to v4.0.94
 
 1.4.0